backend/requirements.txt
Django==5.1.2
djangorestframework==3.15.2
django-cors-headers==4.4.0
pytest==8.3.3
pytest-django==4.9.0
python-dotenv==1.0.1
